Add external config loading, Remove Session, flatten packages.
Updates the SDK adding an external package. This package is resposible
for loading external configuration data such as environment and shared
config files. In addition this would be the customers initial interface
into the SDK loading the SDKs config.

Removes session package from the SDK. This update removes the Session
type and all of its loading logic. This is all replaced with the
external package. The aws.Config type is updated to satisfy the
ConfigProvider interface.

Flattens the client, metadata, request, and credentials packages into
aws.