Feature/config refactor (#14)
Update the v2 SDK removing the `Session` and `ConfigProvider` types. In their place a new package `external` was created to provide external configuration loading. Such as Environment and Shared config.

This change also flattens the SDK's `request`, `client`, `metadata`, and `credentials` packages into `aws.

* Add external config loading, Remove Session, flatten packages. (#8)

Updates the SDK adding an external package. This package is responsible
for loading external configuration data such as environment and shared
config files. In addition this would be the customers initial interface
into the SDK loading the SDKs config.

Removes session package from the SDK. This update removes the Session
type and all of its loading logic. This is all replaced with the
external package. The aws.Config type is updated to satisfy the
ConfigProvider interface.

Flattens the client, metadata, request, and credentials packages into
aws.

* Renames additional parameters of Credentials (#9)

Renames Credentials:

* Valid -> HasKeys
* ProviderName -> Source
* StaticProvider -> StaticCredentialsProvider
* Add CA Bundle injection
* Add EC2 Metadata region provider util
* Add shared config loading tests.

* Correct the load order of ec2 role creds config resolver (#12)

Corrects the external config loading of credentials so that the EC2 role provider is set first as a fallback. Any other credentials set will override it.

* Remove ConfigProvider switching to aws.Config  (#13)

Removing the ConfigProvider interface replacing all usage with `aws.Config`