Refactor endpoints to have a cleaner configuration story (#30)
Updates the endpoints package so that customer configures endpoint
resolution options on the endpoint resolver directlly. This moves
the endpoint resolving options out of Config. This also clarifies
what the SDK will do if the customer provides a custom URL.

Config DisabeSSL, and UseDualStack moved to endpoints.ResolveOptions

Fix #17
Fix #18