service/s3: remove SelectContent until EventStream supported
S3's SelectContent API is not functional in the SDK yet, and was not
supposed to be generated until EventStream support is available. This
change updates the S3 service model using the suppression logic added in
PR #160.

This API will be added back in once EventStream support is available.

Related to #174, #160