Update SDK API operation request send to required Context (#265)
Updates the SDK's API operation request Send method to require a
context.Context value when called. This is done to encourage best
applications to use Context for cancellation and request tracing.
Standardizing on this pattern will also help reduce code paths which
accidentally do not have the Context causing the cancellation and
tracing chain to be lost. Leading to difficult to trace down losses of
cancellation and tracing within an application.

If a context is not desired using context.Background, or
context.TODO instead of specific context, like cancel or trace.

Fix #264