private/model/api: Backfill authtype, STS and Cognito Identity
Backfills the authtype=none modeled trait for STS and Cognito Identity
services. This removes the in code customization for these two services'
APIs that should not be signed.

If the service API models are updated this customization to the code
generation can be removed.

V2 SDK port of: aws/aws-sdk-go#2477