private/model: removes customization for merging services (#454)
The SDK has customizations for merging services like dynamodbstreams with dynamodb; and waf-regional with waf service. This was intended to avoid duplicate shapes. As the SDK moves towards a layout with types, and enums package as seperate package, we need to remove these customizations as they provide little to no value and add complications. This PR removes those unnecessary customizations.