`aws`: Add context on credential providers, EC2Metadata client o… (#461)
This PR makes a breaking change to the aws sdk for go v2, by adding context as a required parameter to each operation on EC2Metadata client. We also modified the Retrieve method on Credential Provider and signer utility to take in a user provided context.

SDK operations that relied on credential providers have been updated to accommodate this change. Tests and examples have also been modified.