Normalize section keys to lowercase
This is to match the behavior of the parsing in Python, and the boto
based AWS CLI, where the keys are normalized to lowercase.

We had a case where things were working using the CLI but when we tried
to do the same action through the Go SDK it didn't work because we had
UPPERCASED the access_key_id.