Refactor SDK's request retry, and error handling behavior
This update refactors the SDK's request retry behavior by updating the
SDK's implementation to a model that will be the standard across the
SDKs. This model simplifies retry behavior, and will retry all
connection errors. Not just write connection reset errors.

This update also refactors the SDK's error wrapping and handling. The
SDK's error types such as awserr.Error implement the Unwrap method
allowing the nested errors to be more easily extracted using the Go 1.13
errors.As, Is, and Unwrap functions.

This update also provides more information about why a request failed,
and was retried, but exhausted attempts. The aws.MaxAttemptsError will wrap
the connection or service API error.