v2 prototype: Move modeled API parameters into types sub package (#542)
Moves the modeled API parameters out of the API client's package and
into a nested package for `types`. With the exception of operation
input/output types, this package will contain all generated
enums, errors and types for an API.

(de)serializers for the generated types will need to go into a third
package that will be depended on by the api client package, (i.e.
`<api>/internal/protocol/<protocol-type>`).