Update signer and related middleware
Updates the signer middleware helper to only add the signer. The compute
SHA256 middleware is broken out into its own header.

Adds a new middleware for setting X-Amz-Content-Sha256 header that will
be paired with the unsigned payload middleware code generation by
default. Other services that need to be customizes for this header can
add the middleware as well, e.g. S3 and Glacier.

Reorders the clock skew middleware to be after retry.