Update map/list query ser to defer key creation
This updates the map/collection query serialization to place the
creation of the entry after the null value check. For lists this
shouldn't make a difference other than readability. For maps this
avoids creating a key that shouldn't be created. This also adds
a short-circuit for empty maps and lists to avoid extra allocations
and work.