Move credential providers to their own module
Moves the credential providers to their own module, `credentials`. This
sets the ground work for the `aws/external` package to be broken out
into its own module, `config` at the root of the repository too.
Breaking out the `credentials` module, allows credential providers to
depend on API clients.

Renamed the `SafeCredentialProvider` to `CredentialsCache`. Also the
credential providers were updated to not embed the credential cache
directly in their implementations. This change allows other credential
caching schemes to be used in the future, such as disk cache, or other
caching strategies. The SDK's external config loading is responsible for
ensuring that the credential providers it resolves are always wrapped
with the credential cache.

Users using the credential provider directly will need to wrap it with
the cache for the credentials to be cached.