codegen: Add flag to direct if SDK is allowed to modify a custom endpoint or not
Adds a new member to the `aws.Endpoint` struct, `HostnameImmutable`. This member
directs if the SDK is allowed to modify the resolved endpoint to meet the
requirements of the API.  If `HostnameImmutable` is true, the SDK will not
attempt to modify the hostname via any customizations, or prefix behavior. Flag
 defaults to false.

- Fixes https://github.com/aws/aws-sdk-go-v2/issues/827 (SDK requiring region for endpoint resolution)
- Fixes https://github.com/aws/aws-sdk-go-v2/issues/328
- Fixes https://github.com/aws/aws-sdk-go-v2/issues/364
- Related to https://github.com/aws/aws-sdk-go-v2/pull/836 When Amazon S3 Accesspoint support is implemented.
- Related to https://github.com/aws/aws-sdk-go-v2/issues/479

Depends on https://github.com/awslabs/smithy-go/pull/222