service/ec2: Fix generation of number and bool struct members to be pointers (#1195)
Adds a backfill customization to the Amazon EC2 API model so that all
unboxed number and boolean shapes are correctly decorated as boxed. The
API does not handle unboxed members. This causes the API client
generated from the model to be unusable in many cases.

The generated API client will contain breaking changes, but these
breaking changes are here to fix the API client that was unusable in
many ways.