service/sqs: Add support for validating message checksums by default
Adds support for the SQS client to automatically validate message
checksums for SendMessage, SendMessageBatch, and ReceiveMessage. This
brings the v2 SDK up to speed with the v1 SDK's behavior.

A DisableMessageChecksumValidation parameter has been added to the
Options struct for SQS package. Setting this to true will disable the
checksum validation. This can be set when creating a client, or per
operation call.