delay WebIdentityRoleOptions validation (#2030)
When loading the default config and `AWS_WEB_IDENTITY_TOKEN_FILE` is set it
triggers the assume role with web identity provider. This provider
requires it's configuration either entirely in environment variables or
shared config. The Go SDK offers an API to configure this provider
further via code which was being ignored or causing early validation
issues. This change allows `AWS_WEB_IDENTITY_TOKEN_FILE` to be set
and all other options to be supplied by Go code relaxing the constraint
that everything has to come from the environment.

fixes aws/aws-sdk-go-v2#2015