Add ApiGatewayExportsNullabilityExceptionIntegration (#2056)
Due to internal model fixes for API Gateway (APIGW) exports, certain shapes
and members that target those shapes that used to have defaults became
nullable. In order to not break the existing Go v2 client API nullability,
this customization adds back default values for affected shapes:

- Root Boolean shapes: false
- Root Number shapes: 0
- Snapshotted Members shapes: inherit defaults

The class of services affected are APIGW services and APIGW services that
migrated to Smithy, seen in `APIGW_NULLABILITY_EXCEPTION_SERVICES`.

A "snapshot" of root-level and member shapes are captured in
`APIGW_exports_nullability_exceptions.json`, a mapping of service shape IDs
to a list of affected root-level and member shapes.

Cases:
- Snapshotted Root level:
  - If a snapshotted root level shape doesn't exist, an exception will be
    thrown.
  - Otherwise, a default will be patched if a default doesn't already exist
- Snapshotted Member level:
  - If a snapshotted member level shape doesn't exist, an exception will be
    thrown.
  - Otherwise, a default will be patched if a default doesn't already exist
- Nonsnapshotted Member level:
  - All nonsnapshotted member level shape that target a snapshotted root level
    will be identified and throw an error.
  - This is prevent breaking changes if a nonsnapshotted member changes target
    from a snapshotted root shape to a nonsnapshotted root shape, as the SDK
    has no prior info on either shape.

Also removes `serviceid.smithy` which was not used anywhere.